Fix tests
Change the name of one of the two tests that had the same name.
Fix the test that wasn't previously run because of the same
name issue: remove the second new line in the end of the config
file because after config is overwritten with a new version, it
contains only one new line at the end.
Add todos for small improvements in some other tests.