Fix python 3.11 deprecation warning
Replace sre_constants.error with re.error, because it's the same
object, but re.error is documented and it is a part of the public API,
whereas sre_constants module is not, and its usage is causing
deprecation warning in python 3.11.
Create custom exception VersionConfigInitializationException
because it isn't obvious why re.error may be thrown when VersionConfig
object is created.